feat(live_connection): refactor , bug fixes
