+++ src/entities/streaming/model/selectors.ts
@@ -2,7 +2,7 @@ import { createSelector } from '@reduxjs/toolkit';
 
 const selectStreaming = (state: RootState) => state.streaming;
 
-export const selectStreamingHistory = createSelector(
+export const selectStreamingSettings = createSelector(
   selectStreaming,
-  streaming => streaming.history,
+  streaming => streaming,
 );
