+++ src/entities/streaming/model/slice.ts
@@ -3,42 +3,33 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 const DEFAULT_HOST = '127.0.0.1';
 const DEFAULT_PORT = '8881';
 
-type History = {
+type StreamingState = {
   ipAddress: string;
   port: string;
   remember: boolean;
 };
 
-type InitialState = {
-  history: History | null;
-};
-
-const initialState: InitialState = {
-  history: {
-    ipAddress: DEFAULT_HOST,
-    port: DEFAULT_PORT,
-    remember: false,
-  },
+const initialState: StreamingState = {
+  ipAddress: DEFAULT_HOST,
+  port: DEFAULT_PORT,
+  remember: false,
 };
 
 const streamingSlice = createSlice({
   name: 'streaming',
   initialState,
   reducers: {
-    setHistory: (state, action: PayloadAction<History>) => {
+    connectionEstabilished: (state, action: PayloadAction<StreamingState>) => {
       const { ipAddress, port, remember } = action.payload;
 
-      state.history = {
+      return {
         ipAddress,
         port,
         remember,
       };
     },
-    clearHistory: state => {
-      state.history = null;
-    },
-    setDefaultHistory: state => {
-      state.history = initialState.history;
+    reset: () => {
+      return initialState;
     },
   },
 });
