+++ src/shared/lib/tcp/index.ts
@@ -1,1 +1,2 @@
 export * from './useTCPSocket';
+export * from './useSendLiveEvent';
