+++ src/shared/lib/tcp/useTCPSocket.ts
@@ -3,10 +3,6 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useToast } from 'react-native-toast-notifications';
 
-import { DrawPoint } from '@entities/drawer';
-import { FlankerWebViewLogRecord } from '@entities/flanker';
-import { StabilityTrackerAnswerValue } from '@shared/api';
-
 import { TCPSocketEmitter } from './TCPSocketEmitter';
 import TCPSocketService from './TCPSocketService';
 
@@ -15,12 +11,6 @@ type Callbacks = {
   onClosed?: () => void;
   onConnected?: () => void;
 };
-
-type LiveEvent =
-  | DrawPoint
-  | StabilityTrackerAnswerValue
-  | FlankerWebViewLogRecord
-  | FlankerWebViewLogRecord[];
 
 export function useTCPSocket(callbacks?: Callbacks) {
   const callbacksRef = useRef(callbacks);
@@ -47,22 +37,6 @@ export function useTCPSocket(callbacks?: Callbacks) {
       TCPSocketService.sendMessage(`${message}$$$`);
     },
     [connected],
-  );
-
-  const sendLiveEvent = useCallback(
-    (data: LiveEvent) => {
-      if (!connected) {
-        return;
-      }
-
-      const liveEvent = {
-        type: 'live_event',
-        data,
-      };
-
-      sendMessage(JSON.stringify(liveEvent));
-    },
-    [sendMessage, connected],
   );
 
   const showDisconnectAlert = useCallback(() => {
@@ -132,7 +106,7 @@ export function useTCPSocket(callbacks?: Callbacks) {
     connecting,
 
     getSocketInfo,
-    sendLiveEvent,
+    sendMessage,
     closeConnection,
   };
 }
